+++ scala3doc/src/dotty/dokka/site/StaticSiteContext.scala
@@ -17,8 +17,7 @@ import org.jetbrains.dokka.plugability.DokkaContext
 import org.jetbrains.dokka.pages.Style
 import org.jetbrains.dokka.model.DisplaySourceSet
 import util.Try
-
-import scala.collection.JavaConverters._
+import collection.JavaConverters._
 
 import dotty.dokka.model.api._
 
@@ -168,7 +167,8 @@ class StaticSiteContext(
     }.toOption
     pathsDri.getOrElse(memberLinkResolver(link).toList)
 
-  def driFor(dest: Path): DRI = DRI(location = s"_.${root.toPath.relativize(dest)}")
+  def driFor(dest: Path): DRI =
+    DRI(location = root.toPath.relativize(dest).iterator.asScala.mkString("."))
 
   def relativePath(myTemplate: LoadedTemplate) = root.toPath.relativize(myTemplate.file.toPath)
